Move the date helper function
