+++ components/EventDetail/EventDetail.js
@@ -3,7 +3,7 @@ import CommentSection from "../CommentSection/CommentSection";
 import DeleteEventButton from "../DeleteEventButton/DeleteEventButton";
 import EditEventButton from "../EditEventButton/EditEventButton";
 import Map from "../Map";
-import { formatDate } from "@/lib/formatDate";
+import { formatDate } from "@/lib/dateHelpers";
 import {
   Card,
   ErrorMessage,
@@ -21,6 +21,7 @@ import { useSession } from "next-auth/react";
 
 import ExpandableText from "./ExpandableText";
 import { locationToString } from "@/lib/fomatLocation";
+import { locationToString } from "@/lib/formatLocation";
 
 export default function EventDetail({ event }) {
   const { data: session } = useSession();
