+++ components/EventPreview/EventPreview.js
@@ -8,7 +8,7 @@ import {
   PreviewTitle,
   PreviewDescription,
 } from "./EventPreview.styled";
-import { formatDate } from "@/lib/formatDate";
+import { formatDate } from "@/lib/dateHelpers";
 
 export default function EventPreview({ event }) {
   const { eventName, shortDescription, start, location, category } = event;
