--- lib/formatDate.js
@@ -1,4 +1,0 @@
-export function formatDate(dateString) {
-  const [year, month, day] = dateString.split("-");
-  return `${day}.${month}.${year}`;
-}
