+++ lib/useEventForm.js
@@ -1,5 +1,6 @@
 import { useState } from "react";
 import { useRouter } from "next/router";
+import { getFormattedTodaysDate } from "./dateHelpers";
 
 export function useEventForm(editEvent) {
   const router = useRouter();
@@ -20,7 +21,7 @@ export function useEventForm(editEvent) {
   // Handle start dates relation to end date and vice versa
 
   const [startDate, setStartDate] = useState(
-    editEvent?.start?.date || getFormattedTodayDate()
+    editEvent?.start?.date || getFormattedTodaysDate()
   );
   const [endDate, setEndDate] = useState(editEvent?.end?.date || startDate);
 
@@ -149,11 +150,4 @@ export const autoGrow = (event) => {
   const textareaContainer = event.target;
   textareaContainer.style.height = "auto";
   textareaContainer.style.height = textareaContainer.scrollHeight + 1 + "px";
-};
-
-export const getFormattedTodayDate = () => {
-  const today = new Date();
-  const month = `0${today.getMonth() + 1}`.slice(-2);
-  const day = `0${today.getDate()}`.slice(-2);
-  return `${today.getFullYear()}-${month}-${day}`;
 };
